fix: TIO-116: split widgets
